Implement click feedback

+++ app/src/main/java/com/yaroslavgorbachh/counter/screen/counter/CounterView.java
@@ -1,25 +1,43 @@
 package com.yaroslavgorbachh.counter.screen.counter;
 
-import android.view.View;
-
+import com.google.android.material.dialog.MaterialAlertDialogBuilder;
 import com.yaroslavgorbachh.counter.R;
 import com.yaroslavgorbachh.counter.data.Models.Counter;
 import com.yaroslavgorbachh.counter.databinding.FragmentCounterBinding;
-import com.yaroslavgorbachh.counter.util.Utility;
+import com.yaroslavgorbachh.counter.feature.Accessibility;
+import com.yaroslavgorbachh.counter.feature.FastCountButton;
 import com.yaroslavgorbachh.counter.util.ViewUtil;
 
 public class CounterView {
+    public interface Callback {
+        void onDelete();
+        void onEdit();
+        void onHistory();
+        void onAbout();
+        void onFullScreen();
+        void onBack();
+        void onInc();
+        void onDec();
+        void onReset();
+    }
 
     private final FragmentCounterBinding mBinding;
     private final Callback mCallback;
 
-    public CounterView(FragmentCounterBinding binding, Callback callback) {
+    public CounterView(FragmentCounterBinding binding, Accessibility accessibility, Callback callback) {
         mBinding = binding;
         mCallback = callback;
 
         binding.toolbar.setOnMenuItemClickListener(i -> {
             if (i.getItemId() == R.id.delete) {
-                callback.onDelete();
+                new MaterialAlertDialogBuilder(mBinding.getRoot().getContext())
+                        .setTitle(mBinding.getRoot().getContext().getString(R.string.deleteCounterDeleteDialog))
+                        .setMessage(R.string.deleteCounterDialogText)
+                        .setPositiveButton(R.string.deleteCounterDialogPositiveButton, (dialog, which) -> {
+                            callback.onDelete();
+                        })
+                        .setNegativeButton(R.string.deleteCounterDialogNegativeButton, null)
+                        .show();
             }
             if (i.getItemId() == R.id.edit) {
                 callback.onEdit();
@@ -39,8 +57,17 @@ public class CounterView {
             return true;
         });
         binding.toolbar.setNavigationOnClickListener(i -> callback.onBack());
-        binding.inc.setOnClickListener(callback::onInc);
-        binding.dec.setOnClickListener(callback::onDec);
+
+        new FastCountButton(binding.inc, () -> {
+            accessibility.playIncFeedback(mBinding.value.getText().toString());
+            mCallback.onInc();
+        }, null);
+
+        new FastCountButton(binding.dec, () -> {
+            accessibility.playDecFeedback(mBinding.value.getText().toString());
+            callback.onDec();
+        }, null);
+
     }
 
     public void setCounter(Counter counter) {
@@ -54,23 +81,4 @@ public class CounterView {
         }
     }
 
-    public interface Callback {
-        void onDelete();
-
-        void onEdit();
-
-        void onHistory();
-
-        void onAbout();
-
-        void onFullScreen();
-
-        void onBack();
-
-        void onInc(View view);
-
-        void onDec(View view);
-
-        void onReset();
-    }
 }
